RFC: postpone runtime configuration to after the rest of the package was initialized
Add a callback system for post-initialization configuration that can be utilize
to spread bits of the configuration throughout the code base but still run them
only in a final stage when importing yt.

Avoid automatically creating a yt configuration file when none is present
