+++ yt/funcs.py
@@ -28,6 +28,7 @@ from packaging.version import Version
 from tqdm import tqdm
 
 from yt._maintenance.deprecation import issue_deprecation_warning
+from yt.config import ytcfg
 from yt.units import YTArray, YTQuantity
 from yt.utilities.exceptions import YTInvalidWidthError
 from yt.utilities.logger import ytLogger as mylog
@@ -188,8 +189,6 @@ def time_execution(func):
         mylog.debug("%s took %0.3f s", func.__name__, (t2 - t1))
         return res
 
-    from yt.config import ytcfg
-
     if ytcfg.get("yt", "time_functions"):
         return wrapper
     else:
@@ -230,7 +229,6 @@ def rootonly(func):
        def some_root_only_function(*args, **kwargs):
            ...
     """
-    from yt.config import ytcfg
 
     @wraps(func)
     def check_parallel_rank(*args, **kwargs):
@@ -343,7 +341,6 @@ def get_pbar(title, maxval):
     and a *maxval*.
     """
     maxval = max(maxval, 1)
-    from yt.config import ytcfg
 
     if (
         ytcfg.get("yt", "suppress_stream_logging")
@@ -361,8 +358,6 @@ def only_on_root(func, *args, **kwargs):
     on the root processor calls the function.  All other processors get "None"
     handed back.
     """
-    from yt.config import ytcfg
-
     if kwargs.pop("global_rootonly", False):
         cfg_option = "global_parallel_rank"
     else:
@@ -379,8 +374,6 @@ def is_root():
     This function returns True if it is on the root processor of the
     topcomm and False otherwise.
     """
-    from yt.config import ytcfg
-
     if not ytcfg.get("yt", "internals", "parallel"):
         return True
     return ytcfg.get("yt", "internals", "topcomm_parallel_rank") == 0
@@ -696,8 +689,6 @@ def parallel_profile(prefix):
     ...     plot = PhasePlot(ds.all_data(), *fields)
     """
     import cProfile
-
-    from yt.config import ytcfg
 
     fn = "%s_%04i_%04i.cprof" % (
         prefix,
