+++ yt/mods.py
@@ -15,7 +15,7 @@ import yt.startup_tasks as __startup_tasks
 from yt import *
 from yt._maintenance.deprecation import issue_deprecation_warning
 from yt.config import ytcfg, ytcfg_defaults
-from yt.utilities.logger import _level
+from yt.utilities.logger import ytLogger
 
 issue_deprecation_warning(
     "The yt.mods module is deprecated.", since="4.1.0", removal="4.2.0"
@@ -24,7 +24,7 @@ issue_deprecation_warning(
 unparsed_args = __startup_tasks.unparsed_args
 
 
-if _level >= int(ytcfg_defaults["yt"]["log_level"]):  # type: ignore
+if ytLogger.getEffectiveLevel() >= int(ytcfg_defaults["yt"]["log_level"]):  # type: ignore
     # This won't get displayed.
     mylog.debug("Turning off NumPy error reporting")
     np.seterr(all="ignore")
