+++ yt/utilities/command_line.py
@@ -1330,6 +1330,7 @@ class YTConfigLocalConfigHandler:
         elif getattr(args, "global", False):
             config_file = global_config_file
         else:
+            config_file: Optional[str] = None
             if local_exists and global_exists:
                 s = (
                     "Yt detected a local and a global configuration file, refusing "
@@ -1347,14 +1348,15 @@ class YTConfigLocalConfigHandler:
                 sys.exit(s)
             elif local_exists:
                 config_file = local_config_file
-            else:
+            elif global_exists:
                 config_file = global_config_file
-            sys.stderr.write(f"INFO: using configuration file: {config_file}.\n")
 
-        if not os.path.exists(config_file):
-            with open(config_file, "w") as f:
-                f.write("[yt]\n")
-
+            if config_file is None:
+                print("WARNING: no configuration file installed.", file=sys.stderr)
+            else:
+                print(
+                    f"INFO: reading configuration file: {config_file}", file=sys.stderr
+                )
         CONFIG.read(config_file)
 
         self.config_file = config_file
@@ -1407,7 +1409,12 @@ class YTConfigSetCmd(YTCommand, YTConfigLocalConfigHandler):
         from yt.utilities.configure import set_config
 
         self.load_config(args)
-
+        if self.config_file is None:
+            self.config_file = os.path.join(os.getcwd(), "yt.toml")
+            print(
+                f"INFO: configuration will be written to {self.config_file}",
+                file=sys.stderr,
+            )
         set_config(args.section, args.option, args.value, self.config_file)
