+++ yt/utilities/logger.py
@@ -1,33 +1,11 @@
 import logging
 import sys
+from typing import Callable, Optional
 
-from yt.config import ytcfg
+from yt.utilities.configure import YTConfig, configuration_callbacks
 
-# This next bit is grabbed from:
-# http://stackoverflow.com/questions/384076/how-can-i-make-the-python-logging-output-to-be-colored
-
-
-def add_coloring_to_emit_ansi(fn):
-    # add methods we need to the class
-    def new(*args):
-        levelno = args[0].levelno
-        if levelno >= 50:
-            color = "\x1b[31m"  # red
-        elif levelno >= 40:
-            color = "\x1b[31m"  # red
-        elif levelno >= 30:
-            color = "\x1b[33m"  # yellow
-        elif levelno >= 20:
-            color = "\x1b[32m"  # green
-        elif levelno >= 10:
-            color = "\x1b[35m"  # pink
-        else:
-            color = "\x1b[0m"  # normal
-        ln = color + args[0].levelname + "\x1b[0m"
-        args[0].levelname = ln
-        return fn(*args)
-
-    return new
+_yt_sh: Optional[logging.StreamHandler] = None
+_original_emitter: Optional[Callable[[logging.LogRecord], None]] = None
 
 
 def set_log_level(level):
@@ -56,14 +34,6 @@ def set_log_level(level):
     ytLogger.setLevel(level)
     ytLogger.debug("Set log level to %s", level)
 
-
-ufstring = "%(name)-3s: [%(levelname)-9s] %(asctime)s %(message)s"
-cfstring = "%(name)-3s: [%(levelname)-18s] %(asctime)s %(message)s"
-
-if ytcfg.get("yt", "stdout_stream_logging"):
-    stream = sys.stdout
-else:
-    stream = sys.stderr
 
 ytLogger = logging.getLogger("yt")
 
@@ -105,6 +75,50 @@ class DeprecatedFieldFilter(logging.Filter):
 
 ytLogger.addFilter(DeprecatedFieldFilter())
 
+# This next bit is grabbed from:
+# http://stackoverflow.com/questions/384076/how-can-i-make-the-python-logging-output-to-be-colored
+
+
+def add_coloring_to_emit_ansi(fn):
+    # add methods we need to the class
+    def new(*args):
+        levelno = args[0].levelno
+        if levelno >= 50:
+            color = "\x1b[31m"  # red
+        elif levelno >= 40:
+            color = "\x1b[31m"  # red
+        elif levelno >= 30:
+            color = "\x1b[33m"  # yellow
+        elif levelno >= 20:
+            color = "\x1b[32m"  # green
+        elif levelno >= 10:
+            color = "\x1b[35m"  # pink
+        else:
+            color = "\x1b[0m"  # normal
+        ln = color + args[0].levelname + "\x1b[0m"
+        args[0].levelname = ln
+        return fn(*args)
+
+    return new
+
+
+ufstring = "%(name)-3s: [%(levelname)-9s] %(asctime)s %(message)s"
+cfstring = "%(name)-3s: [%(levelname)-18s] %(asctime)s %(message)s"
+
+
+def colorize_logging():
+    f = logging.Formatter(cfstring)
+    ytLogger.handlers[0].setFormatter(f)
+    ytLogger.handlers[0].emit = add_coloring_to_emit_ansi(ytLogger.handlers[0].emit)
+
+
+def uncolorize_logging():
+    global _original_emitter, _yt_sh
+    if None not in (_original_emitter, _yt_sh):
+        f = logging.Formatter(ufstring)
+        ytLogger.handlers[0].setFormatter(f)
+        _yt_sh.emit = _original_emitter
+
 
 def disable_stream_logging():
     if len(ytLogger.handlers) > 0:
@@ -113,39 +127,34 @@ def disable_stream_logging():
     ytLogger.addHandler(h)
 
 
-def colorize_logging():
-    f = logging.Formatter(cfstring)
-    ytLogger.handlers[0].setFormatter(f)
-    yt_sh.emit = add_coloring_to_emit_ansi(yt_sh.emit)
+def _runtime_configuration(ytcfg: YTConfig) -> None:
+    # only run this at the end of yt.__init__, after yt.config.ytcfg was instanciated
+
+    global _original_emitter, _yt_sh
+
+    if ytcfg.get("yt", "stdout_stream_logging"):
+        stream = sys.stdout
+    else:
+        stream = sys.stderr
+
+    _level = min(max(ytcfg.get("yt", "log_level"), 0), 50)
+
+    if ytcfg.get("yt", "suppress_stream_logging"):
+        disable_stream_logging()
+    else:
+        _yt_sh = logging.StreamHandler(stream=stream)
+        # create formatter and add it to the handlers
+        formatter = logging.Formatter(ufstring)
+        _yt_sh.setFormatter(formatter)
+        # add the handler to the logger
+        ytLogger.addHandler(_yt_sh)
+        ytLogger.setLevel(_level)
+        ytLogger.propagate = False
+
+        _original_emitter = _yt_sh.emit
+
+        if ytcfg.get("yt", "colored_logs"):
+            colorize_logging()
 
 
-def uncolorize_logging():
-    try:
-        f = logging.Formatter(ufstring)
-        ytLogger.handlers[0].setFormatter(f)
-        yt_sh.emit = original_emitter
-    except NameError:
-        # yt_sh and original_emitter are not defined because
-        # suppress_stream_logging is True, so we continue since there is nothing
-        # to uncolorize
-        pass
-
-
-_level = min(max(ytcfg.get("yt", "log_level"), 0), 50)
-
-if ytcfg.get("yt", "suppress_stream_logging"):
-    disable_stream_logging()
-else:
-    yt_sh = logging.StreamHandler(stream=stream)
-    # create formatter and add it to the handlers
-    formatter = logging.Formatter(ufstring)
-    yt_sh.setFormatter(formatter)
-    # add the handler to the logger
-    ytLogger.addHandler(yt_sh)
-    set_log_level(_level)
-    ytLogger.propagate = False
-
-    original_emitter = yt_sh.emit
-
-    if ytcfg.get("yt", "colored_logs"):
-        colorize_logging()
+configuration_callbacks.append(_runtime_configuration)
