+++ yt/utilities/parallel_tools/parallel_analysis_interface.py
@@ -681,8 +681,6 @@ class CommunicationSystem:
         return new_comm
 
     def _update_parallel_state(self, new_comm):
-        from yt.config import ytcfg
-
         ytcfg["yt", "internals", "topcomm_parallel_size"] = new_comm.size
         ytcfg["yt", "internals", "topcomm_parallel_rank"] = new_comm.rank
         if new_comm.rank > 0 and ytcfg.get("yt", "serialize"):
